+++ FinalProject/Models/StateWage.cs
@@ -8,7 +8,7 @@ namespace FinalProject.Models
     public class StateWage 
     {
         public int ID { get; set; }
-        public double MinWage { get; set; }
+        public decimal MinWage { get; set; }
         public string State { get; set; }
         public DateTime EffectiveDate { get; set; }
     }
