+++ FinalProject/ViewModels/AddCityWageViewModel.cs
@@ -21,7 +21,7 @@ namespace FinalProject.ViewModels
 
         [Required]
         [DataType(DataType.Currency)]
-        public double MinWage { get; set; }
+        public decimal MinWage { get; set; }
 
         [Required]
         [DataType(DataType.Date)]
